C++: move core source files
This renames directories containing the core C++ source code to
lowercase.
