+++ lib/UnoCore/UnoCore.unoproj
@@ -98,22 +98,22 @@
     "cpp/main-mobile.cpp:file",
     "cpp/uno.cpp.uxl:extensions",
     "cpp/uno.h:file",
-    "cpp/Uno/_config.h:file",
-    "cpp/Uno/_internal.h:file",
-    "cpp/Uno/_invoke.cpp:file",
-    "cpp/Uno/Graphics/GLHelper.h:file",
-    "cpp/Uno/Memory.cpp:file",
-    "cpp/Uno/Memory.h:file",
-    "cpp/Uno/NativeStackTrace.cpp:file",
-    "cpp/Uno/ObjectModel.cpp:file",
-    "cpp/Uno/ObjectModel.h:file",
-    "cpp/Uno/ObjectMonitor.h:file",
-    "cpp/Uno/Reflection.cpp:file",
-    "cpp/Uno/Reflection.h:file",
-    "cpp/Uno/Support.cpp:file",
-    "cpp/Uno/Support.h:file",
-    "cpp/Uno/SupportApple.mm:file",
-    "cpp/Uno/Uno.h:file",
+    "cpp/uno/_config.h:file",
+    "cpp/uno/_internal.h:file",
+    "cpp/uno/_invoke.cpp:file",
+    "cpp/uno/graphics/GLHelper.h:file",
+    "cpp/uno/Memory.cpp:file",
+    "cpp/uno/Memory.h:file",
+    "cpp/uno/NativeStackTrace.cpp:file",
+    "cpp/uno/ObjectModel.cpp:file",
+    "cpp/uno/ObjectModel.h:file",
+    "cpp/uno/ObjectMonitor.h:file",
+    "cpp/uno/Reflection.cpp:file",
+    "cpp/uno/Reflection.h:file",
+    "cpp/uno/Support.cpp:file",
+    "cpp/uno/Support.h:file",
+    "cpp/uno/SupportApple.mm:file",
+    "cpp/uno/uno.h:file",
     "cpp/uPlatform/GLHelper.h:file",
     "cpp/uPlatform/GraphicsContext.h:file",
     "cpp/uPlatform/WinAPIHelper.h:file",
