+++ lib/UnoCore/cpp/main-console.cpp
@@ -1,7 +1,7 @@
 // @(MSG_ORIGIN)
 // @(MSG_EDIT_WARNING)
 
-#include <Uno/ObjectModel.h>
+#include <uno/ObjectModel.h>
 @(Main.IncludeDirective)
 @(TypeObjects.Declaration:JoinSorted())
 void uInitRtti(uType*(*factories[])());
