+++ lib/UnoCore/cpp/uPlatform/GraphicsContext.h
@@ -4,7 +4,7 @@
 #pragma once
 #include <XliPlatform/GLContext.h>
 #include <XliPlatform/GL.h>
-#include <Uno/Support.h>
+#include <uno/Support.h>
 
 extern ::Xli::GLContext *_XliGLContextPtr;
