+++ lib/UnoCore/cpp/uno.h
@@ -1,5 +1,5 @@
 // @(MSG_ORIGIN)
 // @(MSG_EDIT_WARNING)
 
-#include <Uno/ObjectModel.h>
-#include <Uno/Support.h>
+#include <uno/ObjectModel.h>
+#include <uno/Support.h>
