+++ lib/UnoCore/cpp/uno/NativeStackTrace.cpp
