+++ lib/UnoCore/cpp/uno/Reflection.cpp
@@ -1,7 +1,7 @@
 // @(MSG_ORIGIN)
 // @(MSG_EDIT_WARNING)
 
-#include <Uno/Reflection.h>
+#include <uno/Reflection.h>
 #include <string>
 #include <unordered_map>
 @{int:IncludeDirective}
