+++ lib/UnoCore/cpp/uno/Reflection.h
@@ -2,7 +2,7 @@
 // @(MSG_EDIT_WARNING)
 
 #pragma once
-#include <Uno/Memory.h>
+#include <uno/Memory.h>
 
 /**
     \addtogroup Reflection
