+++ lib/UnoCore/cpp/uno/Support.cpp
@@ -1,7 +1,7 @@
 // @(MSG_ORIGIN)
 // @(MSG_EDIT_WARNING)
 
-#include <Uno/Support.h>
+#include <uno/Support.h>
 #include <mutex>
 @{byte:IncludeDirective}
