+++ lib/UnoCore/cpp/uno/Support.h
