+++ lib/UnoCore/cpp/uno/_invoke.cpp
@@ -2,7 +2,7 @@
 // @(MSG_EDIT_WARNING)
 
 // Generated by _invoke.py
-#include <Uno/ObjectModel.h>
+#include <uno/ObjectModel.h>
 #define CALL ((void(*)
 #define ME )func)
 #define T void*
