+++ lib/UnoCore/cpp/uno/_invoke.py
@@ -4,7 +4,7 @@ print("""
 // @(MSG_EDIT_WARNING)
 
 // Generated by _invoke.py
-#include <Uno/ObjectModel.h>
+#include <uno/ObjectModel.h>
 #define CALL ((void(*)
 #define ME )func)
 #define T void*
