+++ lib/UnoCore/src/Uno/Threading/Monitor.uno
@@ -3,8 +3,8 @@ using Uno.Compiler.ExportTargetInterop;
 namespace Uno.Threading
 {
     [extern(DOTNET) DotNetType("System.Threading.Monitor")]
-    [extern(CPLUSPLUS) Require("Source.Include", "Uno/ObjectMonitor.h")]
-    [extern(CPLUSPLUS) Require("Source.Include", "Uno/Support.h")]
+    [extern(CPLUSPLUS) Require("Source.Include", "uno/ObjectMonitor.h")]
+    [extern(CPLUSPLUS) Require("Source.Include", "uno/Support.h")]
     public static class Monitor
     {
         public static void Enter(object obj)
